Fixed comparison functions to return proper value.

+++ src/string/wcscmp.c
@@ -51,6 +51,6 @@ wcscmp(const wchar_t *s1, const wchar_t *s2)
 	while (*s1 == *s2++)
 		if (*s1++ == '\0')
 			return (0);
-	/* XXX assumes wchar_t = int */
-	return (*(const unsigned int *)s1 - *(const unsigned int *)--s2);
+	/* XXX assumes wchar_t = short */
+	return (*(const unsigned short *)s1 - *(const unsigned short *)--s2);
 }
