+++ src/string/wcsncmp.c
@@ -46,9 +46,9 @@ wcsncmp(const wchar_t *s1, const wchar_t *s2, size_t n)
 		return (0);
 	do {
 		if (*s1 != *s2++) {
-			/* XXX assumes wchar_t = int */
-			return (*(const unsigned int *)s1 -
-			    *(const unsigned int *)--s2);
+			/* XXX assumes wchar_t = short */
+			return (*(const unsigned short *)s1 -
+			    *(const unsigned short *)--s2);
 		}
 		if (*s1++ == 0)
 			break;
